requirements.txt
docker
Jetson.GPIO
sounddevice
numpy
torch
requests
ollama
pyright